attempting to tune to deal with fact that bright surfaces register
experimenting with increasing erosion kernel size to cut out any extra bright surfaces, which wouldn't work if bright surface is larger than cube, which could very well happen in game. If we raise lower limit on S value, we can cut out more of the bright surfaces, but also lose any side of the cube that is well illuminated.