tuned HSV and kernel sizes
decided that we couldn't pick up both darker and brighter ranges withought lots of noise, current code is geared towards worse-illuminated. Way to workaround this problem is to use a ringlight to have constant lighting, which still leaves the problem of overlap with other well illuminated surfaces. Possible fix would be usage of colored led ring, meaning we can tune for color-tinted cubes. This depends on strength and effect of ring, and must be tested in lab